ArmstrongsPrme
Given two positive integers s and e, where s<=e and find the sum of the armstrong numbers between the given s and e. Then check if the sum of  armstrong numbers is a prime number or not. If it is a prime return true else return false.
PrimeArstrong(2,3) → true
PrimeArstrong(2,7) → false.